# services/model-relay/internal/adminapi/openapi.yaml
openapi: 3.1.0
info:
  title: BiuMind Model-Relay Admin API
  version: "1.0.0"
  description: |
    Admin endpoints for model_relay configuration. Mounted at
    `/v1/admin/*` on the model-relay service (default port 7001).

    Auth: Bearer JWT issued by identity. Permissions are enforced per
    endpoint by `bauth.RoleCache.RequirePermission` middleware. See
    services/identity/migrations/00016_model_relay_perms.sql for the
    permission catalogue.

    All errors share the shape `{"error": {"code", "message", "field?"}}`.
    Stable error codes:
      bad_json, missing_path_param, invalid_uuid, missing_field,
      not_found, conflict, internal,
      credential_in_use, model_in_use,
      upstream_unreachable, upstream_status, upstream_parse,
      upstream_read, reconcile_failed, not_implemented.
servers:
  - url: http://localhost:7001
    description: dev (local)
...
security:
  - bearerAuth: []
paths:
...
  /v1/admin/models:
    get:
      tags: [models]
      summary: List models
      parameters:
        - { in: query, name: status, schema: { type: string } }
        - { in: query, name: family, schema: { type: string } }
        - { in: query, name: min_plan, schema: { type: string, enum: [free, pro, team] } }
        - { in: query, name: q, schema: { type: string } }
      responses:
        "200": { content: { application/json: { schema: { $ref: "#/components/schemas/ModelList" } } }, description: ok }
    post:
      tags: [models]
      summary: Create model (manual_override defaults to true; bound to default group)
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/ModelInput" } } }
      responses:
        "201": { content: { application/json: { schema: { $ref: "#/components/schemas/Model" } } }, description: created }
...
  /v1/admin/models/{id}:
    parameters: [ { $ref: "#/components/parameters/IdPath" } ]
    get:
      tags: [models]
      summary: Get model + bound groups
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  model:  { $ref: "#/components/schemas/Model" }
                  groups: { type: array, items: { $ref: "#/components/schemas/ModelGroup" } }
    patch:
      tags: [models]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/ModelInput" } } } }
      responses:
        "200": { content: { application/json: { schema: { $ref: "#/components/schemas/Model" } } }, description: ok }
    delete:
      tags: [models]
      summary: Cascade-deletes channels + pricing rows for this model.
      responses:
        "204": { description: deleted }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
...
    Capabilities:
      type: object
      properties:
        vision:    { type: boolean }
        tools:     { type: boolean }
        thinking:  { type: boolean }
        cache:     { type: boolean }
        json_mode: { type: boolean }
        audio:     { type: boolean }
    UpstreamRef:
      type: object
      properties:
        vendor_name: { type: string }
        name_rule:   { type: integer }
        source_etag: { type: string }
        synced_at:   { type: string, format: date-time }
    Model:
      type: object
      properties:
        id:               { type: string, format: uuid }
        code:             { type: string }
        display_name:     { type: string }
        family:           { type: string }
        context_window:   { type: integer }
        max_output:       { type: integer }
        capabilities:     { $ref: "#/components/schemas/Capabilities" }
        min_plan:         { type: string, enum: [free, pro, team] }
        status:           { type: string, enum: [active, deprecated, disabled] }
        sort_order:       { type: integer }
        upstream_ref:     { $ref: "#/components/schemas/UpstreamRef" }
        manual_override:  { type: boolean }
        routing_strategy: { type: string, enum: [weighted, lowest_latency, least_busy, lowest_tpm_rpm, cost_aware] }
        is_default_chat:  { type: boolean, description: "admin-designated platform default chat model; at most one model has true" }
        created_at:       { type: string, format: date-time }
        updated_at:       { type: string, format: date-time }
    ModelInput:
      type: object
      required: [code, display_name]
      properties:
        code:             { type: string }
        display_name:     { type: string }
        family:           { type: string }
        context_window:   { type: integer }
        max_output:       { type: integer }
        capabilities:     { $ref: "#/components/schemas/Capabilities" }
        min_plan:         { type: string, enum: [free, pro, team] }
        status:           { type: string, enum: [active, deprecated, disabled] }
        sort_order:       { type: integer }
        routing_strategy: { type: string, enum: [weighted, lowest_latency, least_busy, lowest_tpm_rpm, cost_aware] }
        manual_override:  { type: boolean }
        is_default_chat:  { type: boolean, description: "set true to make this the default chat model (mode must be chat); clears the flag on all other models" }
    ModelList:
      type: object
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Model" } }
        total: { type: integer }

    # ─── Channel ──────────────────────────────────────────────────
...
    ModelGroup:
      type: object
      properties:
        id:          { type: string, format: uuid }
        code:        { type: string }
        name:        { type: string }
        owner_type:  { type: string, enum: [system, org, user] }
        owner_id:    { type: string }
        description: { type: string }
        status:      { type: string, enum: [active, archived] }
        created_at:  { type: string, format: date-time }
        updated_at:  { type: string, format: date-time }

    # ─── Probe / Sync ─────────────────────────────────────────────
...